Move config to theme
Changes config method to declarative so the plugin does not need to depend on shared configuration between the site and plugin.
