+++ lib/delta-theme/components/Menu/ListMenu.astro
@@ -1,6 +1,6 @@
 ---
 import clsx from "clsx";
-import type { MenuItem } from "@config/types";
+import type { MenuItem } from "../../types";
 import Dropdown from "../Dropdown.astro";
 import Icon from "../Icon.astro";
