+++ lib/delta-theme/components/Menu/Menu.astro
@@ -1,5 +1,5 @@
 ---
-import type { MenuItem } from "@config/types";
+import type { MenuItem } from "../../types";
 import type { ComponentProps } from "astro/types";
 import ListMenu from "./ListMenu.astro";
