+++ lib/delta-theme/components/PageLayout.astro
@@ -1,7 +1,7 @@
 ---
 import { getImage } from "astro:assets";
 import { localizedHTML as favicons } from "astro-favicons/middleware";
-import { site } from "@config/site";
+import { config } from "virtual:delta-theme/config";
 import Footer from "./Footer.astro";
 import { Header } from "./Header";
 import "@fontsource/source-sans-pro/400.css";
@@ -16,7 +16,6 @@ interface Props {
 }
 
 const { title, description, image } = Astro.props;
-const { title: siteTitle, googleAnalyticsId } = site;
 const imageUrl =
   !image || typeof image === "string"
     ? image
@@ -26,7 +25,7 @@ const imageUrl =
 <!doctype html>
 <html lang="en">
   <head>
-    {!import.meta.env.DEV && <GoogleAnalytics id={googleAnalyticsId} />}
+    {!import.meta.env.DEV && <GoogleAnalytics id={config.googleAnalyticsId} />}
     <slot name="head-first" />
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
@@ -34,7 +33,7 @@ const imageUrl =
     <link rel="canonical" href={new URL(Astro.url.pathname, Astro.site)} />
     <link rel="sitemap" href="/sitemap-index.xml" />
     <meta name="generator" content={Astro.generator} />
-    <title>{title ? `${title} | ${siteTitle}` : siteTitle}</title>
+    <title>{title ? `${title} | ${config.title}` : config.title}</title>
     {description && <meta name="description" content={description} />}
     <meta
       name="image"
