+++ lib/delta-theme/deltaTheme.ts
@@ -2,27 +2,25 @@ import tailwind from "@astrojs/tailwind";
 import favicons from "astro-favicons";
 import type { AstroIntegration } from "astro";
 import path from "path";
+import { themePlugin } from "./themePlugin";
+import type { ConfigOptions as ThemeConfig } from "./configSchema";
 import { searchPlugin } from "./searchPlugin";
 import { fileURLToPath } from "url";
-
-interface ThemeConfig {
-  /** site name */
-  name: string;
-}
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 export const deltaTheme = (config: ThemeConfig): AstroIntegration[] => {
-  const { name } = config;
+  const { title } = config;
 
   return [
+    themePlugin(config),
     tailwind({
       configFile: path.resolve(__dirname, "./tailwind.config.ts"),
     }),
     searchPlugin(),
     favicons({
-      name,
-      short_name: name,
+      name: title,
+      short_name: title,
       background: "#042436",
       themes: ["#00ADD4"],
     }),
