+++ lib/delta-theme/virtual.d.ts
@@ -1,0 +1,4 @@
+declare module "virtual:delta-theme/config" {
+  import type { ConfigOptions } from "./configSchema";
+  const config: ConfigOptions;
+}
